# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GOEN API",
        "contact": {
            "name": "Jorge Soto",
            "email": "jorgemsm@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/price": {
            "get": {
                "tags": [
                    "Model: Price",
                    "Role: Admin"
                ],
                "summary": "Display a listing of Prices",
                "operationId": "getAllPrices",
                "responses": {
                    "200": {
                        "description": "Prices found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PriceCollection"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Model: Price",
                    "Role: Admin"
                ],
                "summary": "Store price",
                "operationId": "storePrice",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Price's amount",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns created price",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SinglePrice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/price/{price}": {
            "get": {
                "tags": [
                    "Model: Price",
                    "Role: Admin"
                ],
                "summary": "Get a Price by Id",
                "operationId": "getPriceById",
                "parameters": [
                    {
                        "name": "price",
                        "in": "path",
                        "description": "Id of price",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Price found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SinglePrice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Model: Price",
                    "Role: Admin"
                ],
                "summary": "Update price data",
                "operationId": "updatePrice",
                "parameters": [
                    {
                        "name": "price",
                        "in": "path",
                        "description": "Id of price",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Price's new score",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated price",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SinglePrice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Model: Price",
                    "Role: Admin"
                ],
                "summary": "Delete price",
                "operationId": "deletePrice",
                "parameters": [
                    {
                        "name": "price",
                        "in": "path",
                        "description": "Id of price",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Price deleted"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/users/{id}/{role}": {
            "put": {
                "tags": [
                    "Model: Role",
                    "Role: Admin"
                ],
                "summary": "Add Role to specified user",
                "operationId": "addRoleToUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "Role name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role added to User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User already has role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Model: Role",
                    "Role: Admin"
                ],
                "summary": "Remove Role from specified user",
                "operationId": "removeRoleFromUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "Role name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role removed from User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error due to: admins can't change their own roles as admins, user does not have role, user must have at least one role, etc.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Price": {
                "title": "Price",
                "description": "Modules possible registration price",
                "required": [
                    "id",
                    "amount"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Unique identificator",
                        "type": "integer",
                        "format": "int64"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Price",
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "title": "error",
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ResponseMessage": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SinglePrice": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Price"
                    }
                },
                "type": "object"
            },
            "PriceCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Price"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}